+++ test/mock_helper/BrandFilterHelper.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import BrandFilterDTO from '../../../src/data/dto/BrandFilterDTO';
+import BrandFilterDTO from '../../src/data/dto/BrandFilterDTO';
 import BrandHelper from './BrandHelper';
 
 const FIRST_INITIAL_REGEX = /^[ㄱㄲㄴㄷㄸㄹㅁㅂㅃㅅㅆㅇㅈㅉㅊㅋㅍㅌㅎ]$/;
