+++ test/service/BrandService.spec.ts
@@ -4,14 +4,15 @@ import { Done } from 'mocha';
 dotenv.config();
 
 import BrandDTO from '../../src/data/dto/BrandDTO';
-import BrandHelper from '../data/dto/BrandHelper';
 import BrandFilterDTO from '../../src/data/dto/BrandFilterDTO';
-import BrandFilterHelper from '../data/dto/BrandFilterHelper';
 import BrandService from '../../src/service/BrandService';
 import BrandDao from '../../src/dao/BrandDao';
 import PagingDTO from '../../src/data/dto/PagingDTO';
 import ListAndCountDTO from '../../src/data/dto/ListAndCountDTO';
 import { PagingRequestDTO } from '../../src/data/request/common';
+
+import BrandHelper from '../mock_helper/BrandHelper';
+import BrandFilterHelper from '../mock_helper/BrandFilterHelper';
 
 const mockListAndCountDTO: ListAndCountDTO<BrandDTO> =
     new ListAndCountDTO<BrandDTO>(1, [
